+++ changelog.md
@@ -4,6 +4,7 @@
 * Configurable vehicle explosions
 * Removed Multi-Keybindings as it no longer works with Forge anyways
     * Also removed separate keybinds for pull and push
+* Added config to block aircraft in certain dimensions
 
 # 0.5.2